requirements.txt
pyyaml>=6.0
datasets>=2.19.0
transformers>=4.41.0
peft>=0.11.0
accelerate>=0.30.0
torch>=2.2.0
bitsandbytes>=0.43.0
openai>=1.30.0
tqdm>=4.66.0
numpy>=1.26.0
pandas>=2.2.0
human-eval @ git+https://github.com/openai/human-eval.git